+++ src/app.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { LoggerModule } from 'nestjs-pino';
+import { AuthModule } from './auth/auth.module';
 import { validationSchema } from './config/env.validation';
 import { FeedbackModule } from './feedback/feedback.module';
 import { HealthModule } from './health/health.module';
@@ -25,6 +26,7 @@ import { PrismaModule } from './prisma/prisma.module';
     }),
     PrismaModule,
     HealthModule,
+    AuthModule,
     FeedbackModule,
   ],
 })